Do 4 vector muls for ints per loop iteration
Also replaced the fully unrolled tail scalar loop with a loop unrolled by 4.
In the first iteration the code calculates an address and makes a jump so that
only <cnt % unroll_factor> instruction pairs are executed. In the subsequent
iterations the loop operates as a usual loop.

This patch contains critical bugs (presumably in both parts - vectorized and
scalar) so it's impossible to measure performance in the current state.